requirements.txt
Flask
Flask-PyMongo
Flask-WTF
Flask-Login
Werkzeug
python-dotenv
